WIP : GetSamplingList+M2
I am correcting the automatic sampling.
Presently, it is not working (neither in the past it was, actually) because of design problem (not mathematics).

Also, in this version, the Gaussian source correctly works (I did not correct it, it was already working, but OASYS was using an outdated version of LibWiser, probably)